requirements.txt
Django<1.10
django-cms-site
